feat: add user logs to app
