+++ .eslintrc.json
@@ -24,6 +24,7 @@
         "no-underscore-dangle": 0,
         "react/jsx-props-no-spreading": 0,
         "import/no-extraneous-dependencies": 0,
-        "react/jsx-filename-extension": 0
+        "react/jsx-filename-extension": 0,
+        "consistent-return": 0
     }
 }
