--- client/utils/logger/breadCrumb.js
@@ -1,15 +1,0 @@
-import {
-  addBreadcrumb,
-} from '@sentry/react';
-
-export const breadCrumb = (category, message, level, timestamp = Date.now(), data = null) => {
-  addBreadcrumb({
-    category,
-    message,
-    level,
-    timestamp,
-    data,
-  });
-};
-
-export default breadCrumb;
