--- client/utils/logger/error.js
@@ -1,9 +1,0 @@
-import {
-  captureException,
-} from '@sentry/react';
-
-export const error = (err) => {
-  captureException(err);
-};
-
-export default error;
