--- client/utils/logger/event.js
@@ -1,15 +1,0 @@
-import {
-  captureEvent,
-} from '@sentry/react';
-
-export const event = (message, form, ...options) => {
-  captureEvent({
-    message,
-    extra: {
-      form,
-      ...options,
-    },
-  });
-};
-
-export default event;
