--- client/utils/logger/index.js
@@ -1,11 +1,0 @@
-import { breadCrumb } from './breadCrumb';
-import { error } from './error';
-import { event } from './event';
-import { info } from './info';
-
-export {
-  breadCrumb,
-  error,
-  event,
-  info,
-};
