--- client/utils/logger/info.js
@@ -1,9 +1,0 @@
-import {
-  captureMessage,
-} from '@sentry/react';
-
-export const info = (message) => {
-  captureMessage(message);
-};
-
-export default info;
